Add vader test for the gif example
The README.md has an animated gif demonstrating textobj-word-column in
action. This vader test uses roughly the same text and performs the same
actions.

The gif appears to have an error because it uses ciw and then I: to add
colon as a prefix, but since we're already in insert mode, that would
add "I:" as a prefix.

New gif-example test passes.