Support symbols and identical words
Since symbols don't match word boundaries (they're not words), we need
to look for them explicitly. When we're checking for matches, just
return true if the other text is identical.

Changes several functions to take the reference line (where we're coming
from) in addition to the check line (what we're testing for boundaries).

Theoretically, we should test the first and last characters of the column
selection separately so we can look at the start or end boundaries
separately. That would allow us to select ruby/perl variable names that
starts with a symbol.

gif-example test now passes.